requirements.txt
streamlit
langchain
langchain-groq
python-dotenv

openai>=1.12.0